Remove the latest set of warnings (catched by gcc 4). Now everything should be clean.
This commit was SVN r11341.

+++ opal/util/cmd_line.c
@@ -555,7 +555,7 @@ char *opal_cmd_line_get_usage_msg(opal_cmd_line_t *cmd)
             }
             strcat(line, " ");
             for (i = 0; (int)i < option->clo_num_params; ++i) {
-                snprintf(temp, len, "<arg%d> ", i);
+                snprintf(temp, len, "<arg%d> ", (int)i);
                 strcat(line, temp);
             }
             if (option->clo_num_params > 0) {
